requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
sse-starlette==2.0.0

# OpenAI
openai==1.12.0

# Database
mysql-connector-python==8.3.0
sqlalchemy==2.0.25

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Environment & Settings
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0

# CORS
python-multipart==0.0.6
